requirements.txt
urllib3==2.0.3
opencv-python==4.8.0.74
numpy==1.26.0
torch==2.2.0
torchvision==0.17.0
openai==1.12.0
flask==2.3.2
python-dotenv==1.0.0
pandas==2.0.3
ultralytics==8.0.196
httpx==0.26.0